Refactor: move validate date ranges logic to a new file
